+++ language-server/src/tests/hover.rs
@@ -1934,3 +1934,18 @@ pub fn main() {
         find_position_of("_discard")
     );
 }
+
+#[test]
+fn hover_for_custom_type() {
+    assert_hover!(
+        "/// Exciting documentation
+/// Maybe even multiple lines
+type Wibble {
+  /// Some more exciting documentation
+  Wibble(String)
+  /// The most exciting documentation
+  Wobble(arg: Int)
+}",
+        find_position_of("Wibble")
+    );
+}