+++ CHANGELOG.md
@@ -76,6 +76,10 @@
   variable.
   ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
 
+- The language server now shows hover information when hovering over custom type
+  definitions.
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
+
 - The language server is now smarter when producing autocompletions. Imagine
   you're updating your code to fully qualify the uses of the `Json` type: